+++ cli/docs/changelog.md
@@ -1,5 +1,15 @@
 # TBD
 
+# 0.11.4
+### Features
+* Added a `config path` command to print the path to the Kurtosis config file
+* Added a `--force` flag to `config init` to allow users to ignore the interactive confirmation prompt (useful for running in CI)
+* If the user does end up trying to `config init` in a non-interactive shell when the config already exists, throw an error with informative information
+* Added CI checks to ensure that double-init'ing the config without `force` fails, and with `force` succeeds
+
+### Fixes
+* Fixed a bug with determining if the terminal is a TTY
+
 # 0.11.3
 ### Changes
 * Upgraded to the following, which contain dormant Kubernetes code: